[kubectl] set args instead of command
The toolbox provides a tini entrypoint which we do not benefit from
if command is set in the container.
Note that the terminology is slightly different between docker and
kubernetes :

    https://kubernetes.io/docs/tasks/inject-data-application/define-command-argument-container/#notes

The reasons why we need tini inside a container are detailed in :

    https://www.ianlewis.org/en/almighty-pause-container
    https://hackernoon.com/my-process-became-pid-1-and-now-signals-behave-strangely-b05c52cc551c

In a nutshell, sleep does not install a signal handler for SIGTERM.
Outside a container, it is not noticable as the default action will
terminate the sleep process.
Inside the container, when sleep runs with PID 1, the default action
is not taken into account, and overall the SIGTERM sent when the pod
is deleted is discard.

    https://unix.stackexchange.com/a/429595

The process is eventually killed after the default grace period of
30 seconds :

    https://kubernetes.io/docs/concepts/workloads/pods/pod/#termination-of-pods